trains/requirements.txt
Flask>=3.0.2,<3.1.0
factory-boy>=3.3.0,<3.4.0
pytest>=8.0.2,<8.1.0
python-dotenv>=1.0.1,<1.1.0
uWSGI>=2.0.24,<2.1.0
wdb==3.3.0